\name{kernelTDA-package}
\alias{kernelTDA-package}
\alias{kernelTDA}
\docType{package}
\title{
\packageTitle{kernelTDA}
}
\description{
\packageDescription{kernelTDA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{kernelTDA}
\packageIndices{kernelTDA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{kernelTDA}

Maintainer: \packageMaintainer{kernelTDA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
simple examples of the most important functions
}
