% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTDA.R
\docType{package}
\name{kernelTDA}
\alias{kernelTDA}
\title{kernelTDA: Kernels for Persistence Diagrams}
\description{
The kerenelTDA package provides implementation for the most common kernels defined 
in the space of Persistence Diagrams. These kernel can then be used in most kernel based method.
}
