% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Krein_svm.R
\name{Ksvm.default}
\alias{Ksvm.default}
\title{Krein Support Vector Machine}
\usage{
\method{Ksvm}{default}(kernelmat = NULL, y = NULL, cost = 1,
  class.weights = NULL, cross = 0, probability = FALSE,
  fitted = TRUE, subset, ...)
}
\arguments{
\item{kernelmat}{the kernel (Grahm) matrix computed for all observations}

\item{y}{a vector of labels}

\item{cost}{cost of violating the constraint}

\item{class.weights}{a named vector of weights for the different classes, used for asymmetric class sizes. Not all factor levels have to be supplied (default weight: 1). All components have to be named. Specifying "inverse" will choose the weights inversely proportional to the class distribution.}

\item{cross}{number of fold in a k-fold cross validation}

\item{probability}{logical indicating whether the model should allow for probability predictions (default: \code{FALSE}).}

\item{fitted}{logical indicating whether the fitted values should be computed and included in the model or not (default: \code{TRUE})}

\item{subset}{An index vector specifying the cases to be used in the training sample. (NOTE: If given, this argument must be named.)}
}
\description{
Solves a kernelized Support Vector Machine in the case where the kernel used may not be positive semidefinite.
}
