% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence_image.R
\name{pers.image}
\alias{pers.image}
\title{Compute the Persistence Image for a given diagram, using piecewise linear weight functions and Gaussian distribution}
\usage{
pers.image(d1, nbins, dimension, h)
}
\arguments{
\item{d1}{A persistence diagram, in the form of a matrix with 3 columns (first one is the dimension, second is the birth-time, last one is the death-time)}

\item{nbins}{Number of bins for the discretization of the Persistence Surface into the Persistence Image}

\item{dimension}{The dimension of the topological feature (0 for connected components, 1 for cycles etc)}

\item{h}{Standard deviation of the Gaussian}
}
\value{
a \code{nbins}x\code{nbins} matrix containing the Persistence Image
}
\description{
Compute the Persistence Image for a given diagram, using piecewise linear weight functions and Gaussian distribution
}
