% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_distance.R
\name{wasserstein.distance}
\alias{wasserstein.distance}
\title{Compute Lp Wasserstein Distance}
\usage{
wasserstein.distance(d1, d2, q, p = 2)
}
\arguments{
\item{d1}{A persistence diagram (matrix with 3 col where the first one is the dimension, the second is the birth-time and the third is the death-time).}

\item{d2}{A persistence diagram (matrix with 3 col where the first one is the dimension, the second is the birth-time and the third is the death-time).}

\item{q}{internal parameter}

\item{p}{The dimension of the topological feature (0 for connected components, 1 for cycles etc)}
}
\value{
The value for the Lp Wassesterstein computed in (\code{d1}, \code{d2})
}
\description{
Compute Lp Wasserstein Distance
}
